[chore] Add example of how to use StartSession
- Ensures we're using an external package to get accurate and copy-able
examples in our GoDocs
- Extract examples from bugsnag_test
